Update quiver hashcode dependency
Development on the quiver_* subpackages has been discontinued in favour
of the main quiver package at https://github.com/google/quiver-dart. The
code itself is identical, but is more actively maintained.

This migrates to version 2.0.0 of quiver since version 3.0.0 has been
migrated to null-safety, which requires Dart SDK 2.12.0 or Flutter
2.0.0.